requirements.txt
asgiref==3.2.7
astroid==2.4.0
Django==3.0.5
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-tables2==2.3.1
django-tempus-dominus==5.1.2.13
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
wrapt==1.12.1
